\name{./Assg3-package}
\alias{./Assg3-package}
\alias{./Assg3}
\docType{package}
\title{
\packageTitle{./Assg3}
}
\description{
\packageDescription{./Assg3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{./Assg3}
\packageIndices{./Assg3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{./Assg3}

Maintainer: \packageMaintainer{./Assg3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
